commentary-service-> validation request throw jwt

+++ movie-commentary-service/src/main/java/io/javabrains/moviecommentaryservice/controller/CommentController.java
@@ -1,10 +1,9 @@
 package io.javabrains.moviecommentaryservice.controller;
 
 import io.javabrains.moviecommentaryservice.models.Comment;
-import io.javabrains.moviecommentaryservice.security.TokenSubject;
-import io.javabrains.moviecommentaryservice.security.Utils;
+import io.javabrains.moviecommentaryservice.basicSecurity.TokenSubject;
+import io.javabrains.moviecommentaryservice.basicSecurity.Utils;
 import io.javabrains.moviecommentaryservice.service.CommentService;
-import io.jsonwebtoken.Jwts;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,26 +19,26 @@ public class CommentController {
     @PostMapping
     public ResponseEntity<Comment> save(@RequestBody Comment comment, @RequestHeader("Authorization") String token ) {
 
-        TokenSubject tokenSubject=null;
-        if(token!=null){
-             tokenSubject = Utils.getTokenSubjectFromJson(
-                    Jwts.parser()
-                            .setSigningKey(Utils.SECRET)
-                            .parseClaimsJws(token.replace(Utils.TOKEN_PREFIX, ""))
-                            .getBody()
-                            .getSubject()
-            );
-        }
-        System.out.println(tokenSubject);
-        return new ResponseEntity<Comment>(commentService.save(comment), HttpStatus.CREATED);
+        TokenSubject tokenSubject=Utils.validateRequestUsingJWT(token);
+
+        if(tokenSubject!=null)
+            return new ResponseEntity<Comment>(commentService.save(comment), HttpStatus.CREATED);
+        else
+            return new ResponseEntity<Comment>(HttpStatus.UNAUTHORIZED);
     }
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity delete(@PathVariable Long id) {
-        if (commentService.delete(id) >= 1)
-            return new ResponseEntity(HttpStatus.OK);
+    public ResponseEntity delete(@PathVariable Long id, @RequestHeader("Authorization") String token ) {
+        TokenSubject tokenSubject=Utils.validateRequestUsingJWT(token);
+
+        if(tokenSubject!=null) {
+            if (commentService.delete(id) >= 1)
+                return new ResponseEntity(HttpStatus.OK);
+            else
+                return new ResponseEntity(HttpStatus.NO_CONTENT);
+        }
         else
-            return new ResponseEntity(HttpStatus.NO_CONTENT);
+            return new ResponseEntity<Comment>(HttpStatus.UNAUTHORIZED);
     }
